+++ Gruntfile.js
@@ -33,8 +33,7 @@ module.exports = function(grunt) {
       },
       dist: {
         files: {
-          'build/octicons.min.css': 'build/octicons.css',
-          'build/font/octicons.min.css': 'build/font/octicons.css'
+          'build/octicons.min.css': 'build/octicons.css'
         }
       }
     },
@@ -81,45 +80,7 @@ module.exports = function(grunt) {
       }
     },
 
-    webfont: {
-      options: {
-        font: "octicons",
-        fontFamilyName: "Octicons",
-        types: 'eot,woff,woff2,ttf,svg',
-        fontHeight: 96,
-        normalize: false,
-        ascent: 84,
-        descent: 12,
-        htmlDemo: false,
-        codepointsFile: 'lib/font/codepoints.json',
-        templateOptions: {
-          baseClass: 'octicon',
-          classPrefix: 'octicon-',
-          mixinPrefix: 'octicon-',
-          fontFamilyName: "Octicons"
-        }
-      },
-      octicons_css: {
-        src: 'lib/svg/*.svg',
-        dest: 'build/font',
-        options: {
-          template: 'lib/font/template.css'
-        }
-      },
-      octicons_scss: {
-        src: 'lib/svg/*.svg',
-        dest: 'build/font',
-        options: {
-          stylesheet: 'scss',
-          template: 'lib/font/template.scss'
-        }
-      }
-    },
-
     clean: {
-      font: [
-        'build/font/*'
-      ],
       svg: [
         'build/svg/*',
         'build/sprite.octicons.svg',
@@ -131,18 +92,16 @@ module.exports = function(grunt) {
   grunt.loadNpmTasks('grunt-contrib-clean');
   grunt.loadNpmTasks('grunt-postcss');
   grunt.loadNpmTasks('grunt-svg-sprite');
-  grunt.loadNpmTasks('grunt-webfont');
   grunt.loadNpmTasks('grunt-svgmin');
   grunt.loadNpmTasks('grunt-cssnano');
   grunt.loadNpmTasks('grunt-sass');
 
   // build tasks
   grunt.registerTask('css',  ['sass', 'postcss', 'cssnano']);
-  grunt.registerTask('font', ['clean:font', 'webfont']);
   grunt.registerTask('svg', ['clean:svg', 'svgmin', 'svg_sprite']);
 
   // default task, build /dist/
-  grunt.registerTask('default', [ 'svg', 'font', 'css', 'svg_json']);
+  grunt.registerTask('default', [ 'svg', 'css', 'svg_json']);
 
   grunt.registerTask('svg_json', 'create a json object with all minimized svg', function() {
     var result = {}